build system: Avoid tripping over creating a symbolic link.
In 6b0c923d we create a symbolic link instead of copying a
directory. However, in the release tarball, that symlink is resolved
as a copy of the folder to ensure maximum compatibility.  In turn,
that causes re-building a release tarball from a release tarball to
fail due to not being able to create a symlink for a directory that
already exists.

If you wonder why re-building from a release tarball is needed; this
is needed to create the Debian and EL packages.

* doc/Makefile.am: Ignore the failing ‘ln’ command.