web: Implement “read-only” mode for the SPARQL endpoint.
This commit introduces a configuration option to block all
database-altering operations in the Djehuty codebase.  When spawning
multiple instances of Djehuty, this makes reasoning over which
instance can manage the database state, while the others can still
query the database, but not modify it.

With this in place, one can confidently spawn multiple instances to
handle specific tasks in a confined environment (e.g. IIIF, Git
operations).

* doc/running-djehuty.tex: Document “read-only-mode”.
* src/djehuty/web/config.py: Add “sparql_read_only_mode” property.
* src/djehuty/web/database.py: Implement the behavioral changes for
  “sparql_read_only_mode”.
* src/djehuty/web/ui.py: Read the “read-only-mode” configuration option.