web: cache: Use ‘os.fdopen’ instead of ‘open’ to read/write cache files.
Similar to b5f96de37, but then in the
cache module.  When calling ‘os.fdopen’ in a ‘with’ statement, the
file descriptor is closed at the end of the scope.  Therefore, the
call to ‘os.close’ in b5f96de37 attempts to close an already-closed
file descriptor, leading to an error when calculating the MD5 sum of
an uploaded file.

To fix this, I removed the ‘finally’ clause entirely.

* src/djehuty/web/cache.py: Use ‘os.fdopen’ instead of ‘open’ to
  read/write cache files.
* src/djehuty/web/wsgi.py: Remove the ‘finally’ part to avoid closing
  an already-closed file descriptor.